backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.72.0",
    "arize-phoenix>=12.27.0",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "fpdf2>=2.7.0",
    "jigsawstack>=0.3.8",
    "llama-index-callbacks-arize-phoenix>=0.6.1",
    "llama-index-core>=0.14.10",
    "markdown>=3.5.0",
    "openai>=2.7.1",
    "openinference-instrumentation-openai>=0.1.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pdfplumber>=0.11.8",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "tqdm>=4.66.0",
    "uvicorn>=0.40.0",
    "whisper>=1.1.10",
]
